[refactor][jecy77]: FieldGroup 컴포넌트 shared로 이동
